Move OrganizationDomainOwnership to common
