+++ kifi-backend/modules/shoebox/app/com/keepit/model/OrganizationDomainOwnershipRepo.scala
@@ -8,42 +8,8 @@ import com.keepit.common.db.slick.DBSession.{ RWSession, RSession }
 import com.keepit.common.db.slick._
 import com.keepit.common.logging.AccessLog
 import com.keepit.common.time._
-import org.joda.time.DateTime
-import play.api.libs.json._
-import play.api.libs.functional.syntax._
 
 import scala.concurrent.duration.Duration
-
-object OrganizationDomainOwnershipStates extends States[OrganizationDomainOwnership]
-
-case class OrganizationDomainOwnership(
-    id: Option[Id[OrganizationDomainOwnership]] = None,
-    createdAt: DateTime = currentDateTime,
-    updatedAt: DateTime = currentDateTime,
-    state: State[OrganizationDomainOwnership] = OrganizationDomainOwnershipStates.ACTIVE,
-    seq: SequenceNumber[OrganizationDomainOwnership] = SequenceNumber.ZERO,
-    organizationId: Id[Organization],
-    domainId: Id[Domain]) extends ModelWithSeqNumber[OrganizationDomainOwnership] with ModelWithState[OrganizationDomainOwnership] {
-
-  override def withId(id: Id[OrganizationDomainOwnership]): OrganizationDomainOwnership = copy(id = Some(id))
-
-  override def withUpdateTime(now: DateTime): OrganizationDomainOwnership = copy(updatedAt = now)
-
-}
-
-object OrganizationDomainOwnership {
-
-  implicit val format = (
-    (__ \ "id").formatNullable[Id[OrganizationDomainOwnership]] and
-    (__ \ "createdAt").format[DateTime] and
-    (__ \ "updatedAt").format[DateTime] and
-    (__ \ "state").format[State[OrganizationDomainOwnership]] and
-    (__ \ "seq").format[SequenceNumber[OrganizationDomainOwnership]] and
-    (__ \ "organizationId").format[Id[Organization]] and
-    (__ \ "domainId").format[Id[Domain]]
-  )(OrganizationDomainOwnership.apply, unlift(OrganizationDomainOwnership.unapply))
-
-}
 
 @ImplementedBy(classOf[OrganizationDomainOwnershipRepoImpl])
 trait OrganizationDomainOwnershipRepo extends Repo[OrganizationDomainOwnership] with SeqNumberFunction[OrganizationDomainOwnership] {
